Extract package listing and search handlers into handlers/listing.dart
